Add files via upload
This script loads the in the database. 
NOTE: The indexing is not done yet. So, it is possible that one has to manually create an index for the fields that they are working with. Use something like this "dblp.create_index([ ("authors", 1) ])" . This indexes the author's field. Once we know what fields we need to index and what format are going to use (ex: dblp.create_index([ ("authors", "text") ]), we can change our load.py file to perform indexing after loading the data.